move session token exchange point
